+++ src/common/localization/en-us.ts
@@ -129,6 +129,11 @@ export const english: IAppStrings = {
         backEndNotAvailable: "Checkbox feature will work in future version of Form Recognizer service, please stay tuned.",
         addName: "Add a model name...",
         downloadJson: "Download JSON file",
+        errors: {
+            electron: {
+                cantAccessFiles: "Cannot access files in '${folderUri}' for training. Please check if specified folder URI is correct."
+            }
+        }
 
     },
     modelCompose: {
@@ -202,6 +207,14 @@ export const english: IAppStrings = {
         downloadScript: "Analyze with python script",
         defaultLocalFileInput: "Browse for a file...",
         defaultURLInput: "Paste or type URL...",
+        editAndUploadToTrainingSet: "Edit & upload to training set",
+        editAndUploadToTrainingSetNotify: "by clicking on this button, this form will be added to this project, where you can edit these labels.",
+        editAndUploadToTrainingSetNotify2: "We are adding this file to your training set, where you could edit the labels and re-train the model.",
+        uploadInPrgoress: "Upload in progress...",
+        confirmDuplicatedAssetName: {
+            title: "Asset name exists",
+            message: "Asset with name '${name}' exists in project, override?"
+        }
     },
     recentModelsView: {
         header: "Select a model to analyze with",
@@ -328,7 +341,7 @@ export const english: IAppStrings = {
                 browse: "Browse",
                 selectFolder: "Select folder",
                 chooseFolder: "Choose folder",
-                invalidFolderMessage: "Connection [${project.sourceConnection.providerOptions.folderPath}] and project folder [${project.folderPath}] are invalid. Please check the specified folders in the Connection and Project Settings pages",
+                invalidFolderMessage: "Connection [${project.sourceConnection.providerOptions.folderPath}] and/or project folder [${project.folderPath}] are invalid. Please check the specified folders in the Connection and Project Settings pages",
             },
         },
     },
@@ -434,6 +447,8 @@ export const english: IAppStrings = {
                         subIMenuItems: {
                             runOcrOnCurrentDocument: "Run OCR on current document",
                             runOcrOnAllDocuments: "Run OCR on all documents",
+                            runAutoLabelingCurrentDocument: "Run AutoLabeling on current document",
+                            noPredictModelOnProject: "Predict model not avaliable, please train the model first.",
                         }
                     }
                 }